frontend/api/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Diego Sesmas <diegos7167@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
sqlalchemy = "^2.0.20"
asyncpg = "^0.28.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"
argon2-cffi = "^23.1.0"
pydantic-extra-types = "^2.1.0"
httpx-oauth = "^0.13.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
redfin = "^0.1.1"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
